+++ src/main/java/ru/practicum/shareit/user/dao/InMemoryUserDao.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 import ru.practicum.shareit.exception.UserNotFoundException;
 import ru.practicum.shareit.exception.ValidationException;
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -19,8 +19,8 @@ import java.util.Map;
 @Primary
 public class InMemoryUserDao implements UserDao {
 
-    private final HashMap<Integer, User> userMap = new HashMap<>();
-    int generatedId = 0;
+    private final HashMap<Long, User> userMap = new HashMap<>();
+    Long generatedId = 0L;
 
     @Override
     public User createUser(User user) {
@@ -32,7 +32,7 @@ public class InMemoryUserDao implements UserDao {
     }
 
     @Override
-    public User getUserById(int userId) {
+    public User getUserById(Long userId) {
         if (userMap.containsKey(userId)) {
             return userMap.get(userId);
         } else {
@@ -42,7 +42,7 @@ public class InMemoryUserDao implements UserDao {
     }
 
     @Override
-    public User updateUser(int userId, User user) {
+    public User updateUser(Long userId, User user) {
         if (userMap.containsKey(userId)) {
             User mainUser = userMap.get(userId);
 
@@ -67,14 +67,14 @@ public class InMemoryUserDao implements UserDao {
     @Override
     public List<User> getAllUsers() {
         List<User> userList = new ArrayList<>();
-        for (Map.Entry<Integer, User> entry : userMap.entrySet()) {
+        for (Map.Entry<Long, User> entry : userMap.entrySet()) {
             userList.add(entry.getValue());
         }
         return userList;
     }
 
     @Override
-    public void deleteUser(int userId) {
+    public void deleteUser(Long userId) {
         if (userMap.containsKey(userId)) {
             userMap.remove(userId);
             log.info("Удален пользователь с ID: " + userId);
@@ -84,12 +84,12 @@ public class InMemoryUserDao implements UserDao {
         }
     }
 
-    private int generateId() {
+    private Long generateId() {
         return ++generatedId;
     }
 
     private void checkEmail(User user) {
-        for (Map.Entry<Integer, User> entry : userMap.entrySet()) {
+        for (Map.Entry<Long, User> entry : userMap.entrySet()) {
             if (entry.getValue().getEmail().equals(user.getEmail())) {
                 log.warn("Email: " + user.getEmail() + " уже используется другим пользователем");
                 throw new ValidationException("Пользователь с таким email уже зарегистрирован!");
