+++ src/main/java/ru/practicum/shareit/user/dao/UserDao.java
@@ -1,6 +1,6 @@
 package ru.practicum.shareit.user.dao;
 
-import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.model.User;
 
 import java.util.List;
 
@@ -8,11 +8,11 @@ public interface UserDao {
 
     User createUser(User user);
 
-    User getUserById(int userId);
+    User getUserById(Long userId);
 
-    User updateUser(int userId, User user);
+    User updateUser(Long userId, User user);
 
     List<User> getAllUsers();
 
-    void deleteUser(int userId);
+    void deleteUser(Long userId);
 }
